requirements-runtime.txt
pydantic>=2.0
click>=8.0
rich>=13.0
pyyaml>=6.0
python-slugify>=8.0
python-dotenv>=1.0
litellm>=1.40,<2.0
instructor>=1.0,<2.0
httpx>=0.27
fastapi>=0.115
uvicorn[standard]>=0.30
sse-starlette>=2.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
aiosqlite>=0.20
alembic>=1.13
pyjwt>=2.8
bcrypt>=4.0
email-validator>=2.0
